improved on pre-post midnight logic for days when cruiser don't run past "midnight " hour
namely i made sure on the days that the cruiser doesn't run past
midnight I would either (if the time is pre midnight & we can catch the
cruiser and get to our destination pre midnight we'll get the results
back and takes it). if its past midnight then we take the cruiser for
the next best time on that day(if for example its monday 1am then then
the next best time is after 7am.